+++ models/post_easyjson.go
@@ -325,7 +325,7 @@ func easyjson5a72dc82DecodeTparkDbModels4(in *jlexer.Lexer, out *Thread) {
 		}
 		switch key {
 		case "id":
-			out.Id = float32(in.Float32())
+			out.Id = int(in.Int())
 		case "title":
 			out.Title = string(in.String())
 		case "author":
@@ -335,7 +335,7 @@ func easyjson5a72dc82DecodeTparkDbModels4(in *jlexer.Lexer, out *Thread) {
 		case "message":
 			out.Message = string(in.String())
 		case "votes":
-			out.Votes = float32(in.Float32())
+			out.Votes = int(in.Int())
 		case "slug":
 			out.Slug = string(in.String())
 		case "created":
@@ -364,7 +364,7 @@ func easyjson5a72dc82EncodeTparkDbModels4(out *jwriter.Writer, in Thread) {
 		} else {
 			out.RawString(prefix)
 		}
-		out.Float32(float32(in.Id))
+		out.Int(int(in.Id))
 	}
 	{
 		const prefix string = ",\"title\":"
@@ -414,7 +414,7 @@ func easyjson5a72dc82EncodeTparkDbModels4(out *jwriter.Writer, in Thread) {
 		} else {
 			out.RawString(prefix)
 		}
-		out.Float32(float32(in.Votes))
+		out.Int(int(in.Votes))
 	}
 	if in.Slug != "" {
 		const prefix string = ",\"slug\":"
@@ -553,7 +553,7 @@ func easyjson5a72dc82DecodeTparkDbModels5(in *jlexer.Lexer, out *Post) {
 		case "forum":
 			out.Forum = string(in.String())
 		case "thread":
-			out.Thread = float32(in.Float32())
+			out.Thread = int(in.Int())
 		case "created":
 			if data := in.Raw(); in.Ok() {
 				in.AddError((out.Created).UnmarshalJSON(data))
@@ -640,7 +640,7 @@ func easyjson5a72dc82EncodeTparkDbModels5(out *jwriter.Writer, in Post) {
 		} else {
 			out.RawString(prefix)
 		}
-		out.Float32(float32(in.Thread))
+		out.Int(int(in.Thread))
 	}
 	if true {
 		const prefix string = ",\"created\":"
